refactor(ui): consolidate step files and split Step.Body from Section
Reorganizes the steps directory and rebalances responsibilities
between Step.Body and Step.Section:

- The verify-domain wrapper, ProvideEmailStep, and the inner
  email-verification screen all live in a single VerifyDomainStep.tsx
  file. The renamed inner screen is EnterVerificationCodeStep — its
  job is the actual code entry; the wrapper handles the shared header
  and provider routing.
- The configure outer step is consolidated into a single ConfigureStep
  component (no inner wizard yet — it's a single screen for now).
  ConfigureCreateAppStep is gone.
- Step.Body becomes a scrollable container with no padding — pure
  flex+scroll wrapper. Steps that want padded content add a
  Step.Section inside Body. ConfirmationStep and
  TestConfigurationStep adopt the new Step.Body + Step.Section shape.
- ConfigureSSO.tsx updates the wizard tree: outer step ids are
  verify-domain / configure / test / confirmation. The verify-domain
  step renders the new VerifyDomainStep wrapper directly; configure
  step renders ConfigureStep directly (no inline inner wizard).
