+++ packages/ui/src/components/ConfigureSSO/steps/ConfigureStep.tsx
@@ -3,7 +3,7 @@ import { descriptors, Flow, Text } from '@/customizables';
 import { Step } from '../elements/Step';
 import { useWizard } from '../elements/Wizard';
 
-export const ConfigureCreateApp = (): JSX.Element => {
+export const ConfigureStep = (): JSX.Element => {
   const { goNext, goPrev, isFirstStep, isLastStep } = useWizard();
 
   return (
@@ -14,11 +14,13 @@ export const ConfigureCreateApp = (): JSX.Element => {
       >
         <Step.Header
           title='Configure Okta Workforce'
-          description='Create a new enterprise application in your Okta Dashboard.'
+          description='Create a new enterprise application in your Okta Dashboard'
         />
 
         <Step.Body>
-          <Text>UI goes here</Text>
+          <Step.Section>
+            <Text>UI goes here</Text>
+          </Step.Section>
         </Step.Body>
 
         <Step.Footer>
