+++ packages/ui/src/components/ConfigureSSO/steps/TestConfigurationStep.tsx
@@ -18,7 +18,9 @@ export const TestConfigurationStep = (): JSX.Element => {
         />
 
         <Step.Body>
-          <Text>UI goes here</Text>
+          <Step.Section>
+            <Text>UI goes here</Text>
+          </Step.Section>
         </Step.Body>
 
         <Step.Footer>
