+++ packages/ui/src/components/ConfigureSSO/steps/index.ts
@@ -1,6 +1,4 @@
-export { ConfigureCreateApp } from './ConfigureCreateAppStep';
+export { ConfigureStep } from './ConfigureStep';
 export { ConfirmationStep } from './ConfirmationStep';
-export { ProvideEmail } from './ProvideEmailStep';
 export { TestConfigurationStep } from './TestConfigurationStep';
-export { VerifyEmailAddressStep } from './VerifyEmailAddressStep';
-export { VerifyEmailDomainStep } from './VerifyEmailDomainStep';
+export { VerifyDomainStep } from './VerifyDomainStep';
